requirements.txt
fastapi==0.118.0        # a recent version known to support Pydantic v2
uvicorn[standard]==0.24.0  # this is fine
python-multipart==0.0.6     # okay
pydantic==2.5.0            # you can stick with 2.x if everything supports it
pydantic-settings==2.1.0   # matches Pydantic 2.x line

google-genai==1.49.0       # or at least the latest in the 1.x line; PyPI shows 1.49.0 as of Nov 2025 :contentReference[oaicite:10]{index=10}

paddleocr == 3.1.0    # if you prefer staying on 2.x line; or go to 3.x with matching paddlepaddle
paddlepaddle == 3.1.0 

# Computer Vision
deepface==0.0.79
tensorflow==2.20.0


numpy==1.24.3
scipy==1.11.4

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

sqlalchemy==2.0.23
alembic==1.13.0

python-dotenv==1.0.0
aiofiles==23.2.1

pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

python-json-logger==2.0.7
